requirements.txt
Flask==2.2.2
WTForms==3.0.1
python-dotenv==0.21.0
Werkzeug==2.2.2